+++ IsekaiMod/Content/ContentAdder.cs
@@ -60,11 +60,7 @@ namespace IsekaiMod.Content
                 Features.IsekaiProtagonist.CharacterDevelopment.AlphaStrike.Add();
                 Features.IsekaiProtagonist.CharacterDevelopment.BetaStrike.Add();
                 Features.IsekaiProtagonist.CharacterDevelopment.GammaStrike.Add();
-                Features.IsekaiProtagonist.CharacterDevelopment.AcidImmunity.Add();
-                Features.IsekaiProtagonist.CharacterDevelopment.ColdImmunity.Add();
-                Features.IsekaiProtagonist.CharacterDevelopment.ElectricityImmunity.Add();
-                Features.IsekaiProtagonist.CharacterDevelopment.FireImmunity.Add();
-                Features.IsekaiProtagonist.CharacterDevelopment.SonicImmunity.Add();
+                Features.IsekaiProtagonist.CharacterDevelopment.EnergyImmunity.Add();
                 Features.IsekaiProtagonist.CharacterDevelopment.ExceptionalSummoningFeature.Add();
                 Features.IsekaiProtagonist.CharacterDevelopment.CharacterDevelopmentSelection.Add();
                 Features.IsekaiProtagonist.OverpoweredAbility.AutoEmpowerFeature.Add();